# hfitd/docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact:
    email: daniele@smartango.com
    name: Daniele Cruciani
  description: API for managing database fixtures, file operations, and package templates
  license:
    name: MIT
    url: https://github.com/danielecr/hot-fixture-tool/blob/main/LICENSE
  termsOfService: https://github.com/danielecr/hot-fixture-tool
  title: Hot Fixture Tool Daemon API
  version: 1.0.0
paths:
...
  /packdownload/{templatename}:
    post:
      consumes:
      - application/json
      description: Download a rendered package template with parameters
      parameters:
      - description: Template name
        example: webapp-starter
        in: path
        name: templatename
        required: true
        type: string
      - description: Template parameters
        in: body
        name: parameters
        required: true
        schema:
          type: object
      produces:
      - application/gzip
      responses:
        "200":
          description: Template package file
          schema:
            type: file
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Template not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download package template
      tags:
      - templates
  /packtmpl:
    get:
      consumes:
      - application/json
      description: Get list of all package templates for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of template objects
          schema:
            items:
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List package templates
      tags:
      - templates
  /packtmpl/{templatename}:
    delete:
      consumes:
      - application/json
      description: Delete a specific package template
      parameters:
      - description: Template name
        example: webapp-starter
        in: path
        name: templatename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Template not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete package template
      tags:
      - templates
    get:
      consumes:
      - application/json
      description: Get a specific package template by name
      parameters:
      - description: Template name
        example: webapp-starter
        in: path
        name: templatename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Template object
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Template not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get package template
      tags:
      - templates
    post:
      consumes:
      - application/json
      description: Create or update a package template
      parameters:
      - description: Template name
        example: webapp-starter
        in: path
        name: templatename
        required: true
        type: string
      - description: Template data
        in: body
        name: template
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create/Update package template
      tags:
      - templates
  /packtmplpackupld:
    post:
      consumes:
      - multipart/form-data
      description: Upload multiple package templates in a single request
      parameters:
      - description: Templates archive file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Bulk upload package templates
      tags:
      - templates
...
securityDefinitions:
  BearerAuth:
    description: JWT Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"